6-size.c

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-printf("Size of char: %lu bytes\n", sizeof(char));
-printf("Size of int: %lu bytes\n", sizeof(int));
-printf("Size of long: %lu bytes\n", sizeof(long));
-printf("Size of long long: %lu bytes\n", sizeof(long long));
-printf("Size of float: %lu bytes\n", sizeof(float));
+printf("Size of char: %lu bytes\n", (unsigned long)sizeof(char));
+printf("Size of int: %lu bytes\n", (unsigned long)sizeof(int));
+printf("Size of long: %lu bytes\n", (unsigned long)sizeof(long));
+printf("Size of long long: %lu bytes\n", (unsigned long)sizeof(long long));
+printf("Size of float: %lu bytes\n", (unsigned long)sizeof(float));
 return (0);
 }
